transition Quotes component from options to composition API
